Field4D_Site/backend_fastapi/requirements.txt
fastapi
uvicorn[standard]
google-cloud-bigquery
pydantic-settings
python-dotenv
httpx
